+++ src/table.jl
@@ -7,3 +7,14 @@ Tables.columns(g::SpaceLiDAR.Granule) = Tables.CopiedColumns(joinpartitions(g))
 Tables.istable(::Type{<:SpaceLiDAR.ICESat_Granule}) = true
 Tables.columnaccess(::Type{<:SpaceLiDAR.ICESat_Granule}) = true
 Tables.columns(g::SpaceLiDAR.ICESat_Granule) = points(g)
+
+function materialize!(df::DataFrame)
+    for (name, col) in zip(names(df), eachcol(df))
+        if col isa CategoricalArray
+            df[!, name] = String.(col)
+        elseif col isa Fill
+            df[!, name] = Vector(col)
+        end
+    end
+    df
+end